+++ git_machete/cli.py
@@ -15,7 +15,7 @@ from git_machete.constants import EscapeCodes
 from git_machete.docs import short_docs, long_docs
 from git_machete.exceptions import MacheteException, StopInteraction
 from git_machete.git_operations import AnyRevision, GitContext, LocalBranchShortName, RemoteBranchShortName
-from git_machete.utils import fmt, perform_interactive_slide_out, underline, excluding, warn
+from git_machete.utils import fmt, underline, excluding, warn
 
 T = TypeVar('T')
 
@@ -36,7 +36,7 @@ command_groups: List[Tuple[str, List[str]]] = [
     ("Build, display and modify the tree of branch dependencies",
      ["add", "anno", "discover", "edit", "status"]),
     ("List, check out and delete branches",
-     ["delete-unmanaged", "go", "is-managed", "list", "show"]),
+     ["clean", "delete-unmanaged", "go", "is-managed", "list", "show"]),
     ("Determine changes specific to the given branch",
      ["diff", "fork-point", "log"]),
     ("Update git history in accordance with the tree of branch dependencies",
@@ -156,6 +156,10 @@ def create_cli_parser() -> argparse.ArgumentParser:
     anno_parser.add_argument(
         '-H', '--sync-github-prs', action='store_true', default=argparse.SUPPRESS)
 
+    clean_parser = subparsers.add_parser('clean', usage=argparse.SUPPRESS, add_help=False, parents=[common_args_parser])
+    clean_parser.add_argument('-H', '--checkout-my-github-prs', action='store_true', default=argparse.SUPPRESS)
+    clean_parser.add_argument('-y', '--yes', action='store_true', default=argparse.SUPPRESS)
+
     delete_unmanaged_parser = subparsers.add_parser(
         'delete-unmanaged',
         usage=argparse.SUPPRESS,
@@ -221,7 +225,7 @@ def create_cli_parser() -> argparse.ArgumentParser:
         usage=argparse.SUPPRESS,
         add_help=False,
         parents=[common_args_parser])
-    github_parser.add_argument('subcommand', choices=['anno-prs', 'checkout-prs', 'create-pr', 'retarget-pr'])
+    github_parser.add_argument('subcommand', choices=['anno-prs', 'checkout-prs', 'create-pr', 'retarget-pr', 'sync'])
     github_parser.add_argument('pr_no', nargs='*', type=int)
     github_parser.add_argument('--all', action='store_true')
     github_parser.add_argument('--by')
@@ -296,6 +300,7 @@ def create_cli_parser() -> argparse.ArgumentParser:
         parents=[common_args_parser])
     slide_out_parser.add_argument('branches', nargs='*')
     slide_out_parser.add_argument('-d', '--down-fork-point')
+    slide_out_parser.add_argument('--delete', action='store_true')
     slide_out_parser.add_argument('-M', '--merge', action='store_true')
     slide_out_parser.add_argument('-n', action='store_true')
     slide_out_parser.add_argument('--no-edit-merge', action='store_true')
@@ -379,6 +384,8 @@ def update_cli_opts_using_parsed_args(
             cli_opts.opt_checked_out_since = arg
         elif opt == "color":
             cli_opts.opt_color = arg
+        elif opt == "delete":
+            cli_opts.opt_delete = True
         elif opt == "down_fork_point":
             cli_opts.opt_down_fork_point = AnyRevision.of(arg)
         elif opt == "debug":
@@ -531,9 +538,9 @@ def launch(orig_args: List[str]) -> None:
                     f"{machete_client.definition_file_path} is a directory "
                     "rather than a regular file, aborting")
 
-        perform_interactive_slide_out_flag = perform_interactive_slide_out(cmd)
+        should_perform_interactive_slide_out = MacheteClient.should_perform_interactive_slide_out(cmd)
         if cmd == "add":
-            machete_client.read_definition_file(perform_interactive_slide_out=perform_interactive_slide_out_flag)
+            machete_client.read_definition_file(perform_interactive_slide_out=should_perform_interactive_slide_out)
             branch = get_branch_arg_or_current_branch(cli_opts, git)
             machete_client.add(
                 branch=branch,
@@ -541,13 +548,13 @@ def launch(orig_args: List[str]) -> None:
                 opt_as_root=cli_opts.opt_as_root,
                 opt_yes=cli_opts.opt_yes)
         elif cmd == "advance":
-            machete_client.read_definition_file(perform_interactive_slide_out=perform_interactive_slide_out_flag)
+            machete_client.read_definition_file(perform_interactive_slide_out=should_perform_interactive_slide_out)
             git.expect_no_operation_in_progress()
             current_branch = git.get_current_branch()
             machete_client.expect_in_managed_branches(current_branch)
             machete_client.advance(branch=current_branch, opt_yes=cli_opts.opt_yes)
         elif cmd == "anno":
-            machete_client.read_definition_file(perform_interactive_slide_out=perform_interactive_slide_out_flag,
+            machete_client.read_definition_file(perform_interactive_slide_out=should_perform_interactive_slide_out,
                                                 verify_branches=False)
             if cli_opts.opt_sync_github_prs:
                 machete_client.sync_annotations_to_github_prs()
@@ -558,11 +565,18 @@ def launch(orig_args: List[str]) -> None:
                     machete_client.annotate(branch, parsed_cli.annotation_text)
                 else:
                     machete_client.print_annotation(branch)
+        elif cmd == "clean":
+            machete_client.read_definition_file(perform_interactive_slide_out=should_perform_interactive_slide_out)
+            if 'checkout_my_github_prs' in parsed_cli:
+                machete_client.checkout_github_prs(pr_nos=[],
+                                                   my_opened_prs=True)
+            machete_client.delete_unmanaged(opt_yes=cli_opts.opt_yes)
+            machete_client.delete_untracked(opt_yes=cli_opts.opt_yes)
         elif cmd == "delete-unmanaged":
-            machete_client.read_definition_file(perform_interactive_slide_out=perform_interactive_slide_out_flag)
+            machete_client.read_definition_file(perform_interactive_slide_out=should_perform_interactive_slide_out)
             machete_client.delete_unmanaged(opt_yes=cli_opts.opt_yes)
         elif cmd in {"diff", alias_by_command["diff"]}:
-            machete_client.read_definition_file(perform_interactive_slide_out=perform_interactive_slide_out_flag)
+            machete_client.read_definition_file(perform_interactive_slide_out=should_perform_interactive_slide_out)
             machete_client.diff(branch=parsed_cli.branch, opt_stat=cli_opts.opt_stat)  # passing None if not specified
         elif cmd == "discover":
             # No need to read definition file.
@@ -578,7 +592,7 @@ def launch(orig_args: List[str]) -> None:
             # No need to read definition file.
             print(os.path.abspath(machete_client.definition_file_path))
         elif cmd == "fork-point":
-            machete_client.read_definition_file(perform_interactive_slide_out=perform_interactive_slide_out_flag)
+            machete_client.read_definition_file(perform_interactive_slide_out=should_perform_interactive_slide_out)
             branch = get_branch_arg_or_current_branch(cli_opts, git)
             if cli_opts.opt_inferred:
                 print(machete_client.fork_point(
@@ -608,7 +622,7 @@ def launch(orig_args: List[str]) -> None:
                     use_overrides=True,
                     opt_no_detect_squash_merges=cli_opts.opt_no_detect_squash_merges))
         elif cmd in {"go", alias_by_command["go"]}:
-            machete_client.read_definition_file(perform_interactive_slide_out=perform_interactive_slide_out_flag)
+            machete_client.read_definition_file(perform_interactive_slide_out=should_perform_interactive_slide_out)
             git.expect_no_operation_in_progress()
             current_branch = git.get_current_branch()
             dest = machete_client.parse_direction(
@@ -617,7 +631,7 @@ def launch(orig_args: List[str]) -> None:
                 git.checkout(dest)
         elif cmd == "github":
             github_subcommand = parsed_cli.subcommand
-            machete_client.read_definition_file(perform_interactive_slide_out=perform_interactive_slide_out_flag)
+            machete_client.read_definition_file(perform_interactive_slide_out=should_perform_interactive_slide_out)
 
             if 'draft' in parsed_cli and github_subcommand != 'create-pr':
                 raise MacheteException("'--draft' option is only valid with 'create-pr' subcommand.")
@@ -636,7 +650,8 @@ def launch(orig_args: List[str]) -> None:
                 machete_client.checkout_github_prs(pr_nos=parsed_cli.pr_no if 'pr_no' in parsed_cli else [],
                                                    all_opened_prs=parsed_cli.all if 'all' in parsed_cli else False,
                                                    my_opened_prs=parsed_cli.mine if 'mine' in parsed_cli else False,
-                                                   opened_by=parsed_cli.by if 'by' in parsed_cli else None)
+                                                   opened_by=parsed_cli.by if 'by' in parsed_cli else None,
+                                                   fail_on_missing_current_user_for_my_opened_prs=False)
             elif github_subcommand == "create-pr":
                 current_branch = git.get_current_branch()
                 machete_client.create_github_pr(
@@ -647,8 +662,13 @@ def launch(orig_args: List[str]) -> None:
                 current_branch = git.get_current_branch()
                 machete_client.expect_in_managed_branches(current_branch)
                 machete_client.retarget_github_pr(current_branch)
+            elif github_subcommand == "sync":
+                machete_client.checkout_github_prs(pr_nos=[],
+                                                   my_opened_prs=True)
+                machete_client.delete_unmanaged(opt_yes=False)
+                machete_client.delete_untracked(opt_yes=cli_opts.opt_yes)
         elif cmd == "is-managed":
-            machete_client.read_definition_file(perform_interactive_slide_out=perform_interactive_slide_out_flag)
+            machete_client.read_definition_file(perform_interactive_slide_out=should_perform_interactive_slide_out)
             branch = get_branch_arg_or_current_branch(cli_opts, git)
             if branch is None or branch not in machete_client.managed_branches:
                 exit_script(1)
@@ -661,7 +681,7 @@ def launch(orig_args: List[str]) -> None:
                 raise MacheteException(
                     f"`git machete list {category}` does not expect extra arguments")
 
-            machete_client.read_definition_file(perform_interactive_slide_out=perform_interactive_slide_out_flag)
+            machete_client.read_definition_file(perform_interactive_slide_out=should_perform_interactive_slide_out)
             res = []
             if category == "addable":
                 def strip_remote_name(remote_branch: RemoteBranchShortName) -> LocalBranchShortName:
@@ -692,11 +712,11 @@ def launch(orig_args: List[str]) -> None:
             if res:
                 print("\n".join(res))
         elif cmd in {"log", alias_by_command["log"]}:
-            machete_client.read_definition_file(perform_interactive_slide_out=perform_interactive_slide_out_flag)
+            machete_client.read_definition_file(perform_interactive_slide_out=should_perform_interactive_slide_out)
             branch = get_branch_arg_or_current_branch(cli_opts, git)
             machete_client.log(branch)
         elif cmd == "reapply":
-            machete_client.read_definition_file(perform_interactive_slide_out=perform_interactive_slide_out_flag)
+            machete_client.read_definition_file(perform_interactive_slide_out=should_perform_interactive_slide_out)
             git.expect_no_operation_in_progress()
             current_branch = git.get_current_branch()
             if "fork_point" in parsed_cli:
@@ -718,7 +738,7 @@ def launch(orig_args: List[str]) -> None:
                         '`show current` with a <branch> argument does not make sense')
                 print(branch)
             else:
-                machete_client.read_definition_file(perform_interactive_slide_out=perform_interactive_slide_out_flag,
+                machete_client.read_definition_file(perform_interactive_slide_out=should_perform_interactive_slide_out,
                                                     verify_branches=False)
                 print(
                     '\n'.join(machete_client.parse_direction(
@@ -729,17 +749,18 @@ def launch(orig_args: List[str]) -> None:
                     ))
                 )
         elif cmd == "slide-out":
-            machete_client.read_definition_file(perform_interactive_slide_out=perform_interactive_slide_out_flag)
+            machete_client.read_definition_file(perform_interactive_slide_out=should_perform_interactive_slide_out)
             git.expect_no_operation_in_progress()
             branches = parsed_cli_as_dict.get('branches', [git.get_current_branch()])
             machete_client.slide_out(
                 branches_to_slide_out=list(map(LocalBranchShortName.of, branches)),
+                opt_delete=cli_opts.opt_delete,
                 opt_down_fork_point=cli_opts.opt_down_fork_point,
                 opt_merge=cli_opts.opt_merge,
                 opt_no_interactive_rebase=cli_opts.opt_no_interactive_rebase,
                 opt_no_edit_merge=cli_opts.opt_no_edit_merge)
         elif cmd == "squash":
-            machete_client.read_definition_file(perform_interactive_slide_out=perform_interactive_slide_out_flag)
+            machete_client.read_definition_file(perform_interactive_slide_out=should_perform_interactive_slide_out)
             git.expect_no_operation_in_progress()
             current_branch = git.get_current_branch()
             if "fork_point" in parsed_cli:
@@ -752,7 +773,7 @@ def launch(orig_args: List[str]) -> None:
                     use_overrides=True,
                     opt_no_detect_squash_merges=cli_opts.opt_no_detect_squash_merges))
         elif cmd in {"status", alias_by_command["status"]}:
-            machete_client.read_definition_file(perform_interactive_slide_out=perform_interactive_slide_out_flag)
+            machete_client.read_definition_file(perform_interactive_slide_out=should_perform_interactive_slide_out)
             machete_client.expect_at_least_one_managed_branch()
             machete_client.status(
                 warn_on_yellow_edges=True,
@@ -768,7 +789,7 @@ def launch(orig_args: List[str]) -> None:
                 raise MacheteException(
                     "Invalid argument for `--return-to`. "
                     "Valid arguments: `here|nearest-remaining|stay`.")
-            machete_client.read_definition_file(perform_interactive_slide_out=perform_interactive_slide_out_flag)
+            machete_client.read_definition_file(perform_interactive_slide_out=should_perform_interactive_slide_out)
             git.expect_no_operation_in_progress()
             machete_client.traverse(
                 opt_fetch=cli_opts.opt_fetch,
@@ -783,7 +804,7 @@ def launch(orig_args: List[str]) -> None:
                 opt_start_from=cli_opts.opt_start_from,
                 opt_yes=cli_opts.opt_yes)
         elif cmd == "update":
-            machete_client.read_definition_file(perform_interactive_slide_out=perform_interactive_slide_out_flag)
+            machete_client.read_definition_file(perform_interactive_slide_out=should_perform_interactive_slide_out)
             git.expect_no_operation_in_progress()
             if "fork_point" in parsed_cli:
                 machete_client.check_that_forkpoint_is_ancestor_or_equal_to_tip_of_branch(