+++ git_machete/options.py
@@ -12,6 +12,7 @@ class CommandLineOptions:
         self.opt_checked_out_since: Optional[str] = None
         self.opt_color: str = "auto"
         self.opt_debug: bool = False
+        self.opt_delete: bool = False
         self.opt_down_fork_point: Optional[AnyRevision] = None
         self.opt_draft: bool = False
         self.opt_fetch: bool = False