+++ RELEASE_NOTES.md
@@ -5,6 +5,8 @@
 
 - added: `--all`, `--mine`, `--by` flags and parameter `<PR-number-1> ... <PR-number-N>` to `git machete github checkout-prs`
 - fixed: cherry-pick/merge/rebase/revert is detected on a per-worktree basis
+- added: command `git machete clean` with `--checkout-my-github-prs` flag and its equivalent `git machete github sync`
+- added: `--delete` flag to `git machete slide-out` command for deleting slid-out branches from git
 
 ## New in git-machete 3.7.2